media-indexer/worker/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-indexer-worker"
version = "2.0.0"
description = "Background scan/index worker for MKLanLocal v2.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "huggingface_hub>=0.32.0",
  "numpy>=2.1.0",
  "onnxruntime>=1.20.0",
  "opencv-contrib-python-headless>=4.10.0.84",
  "Pillow>=11.0.0",
  "pgvector>=0.3.5",
  "psycopg[binary]>=3.2.0",
  "pytesseract>=0.3.13",
  "sqlalchemy>=2.0.36",
  "torch>=2.5.0",
  "transformers>=4.46.0",
  "uvicorn[standard]>=0.32.0",
  "watchdog>=6.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
